requirements.txt
streamlit
graphviz
pymongo
python-dotenv
pandas
networkx